+++ core/src/commonMain/kotlin/zakadabar/stack/data/schema/UuidValidationRuleList.kt
@@ -14,24 +14,36 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package zakadabar.stack.frontend.builtin.form.fields
+package zakadabar.stack.data.schema
 
-import zakadabar.stack.data.DtoBase
-import zakadabar.stack.frontend.builtin.form.ZkForm
+import zakadabar.stack.data.builtin.misc.Secret
+import zakadabar.stack.util.PublicApi
+import zakadabar.stack.util.UUID
 import kotlin.reflect.KMutableProperty0
 
-open class ZkStringField<T : DtoBase>(
-    form: ZkForm<T>,
-    prop: KMutableProperty0<String>
-) : ZkStringBase<T, String>(
-    form = form,
-    prop = prop
-) {
+class UuidValidationRuleList(val kProperty: KMutableProperty0<UUID>) : ValidationRuleList<UUID> {
 
-    override fun getPropValue() = prop.get()
+    var defaultValue = UUID.NIL
 
-    override fun setPropValue(value: String) {
-        prop.set(value)
+    private val rules = mutableListOf<ValidationRule<UUID>>()
+
+    override fun validate(report: ValidityReport) {
+        val value = kProperty.get()
+        for (rule in rules) {
+            rule.validate(value, report)
+        }
     }
+
+    @PublicApi
+    infix fun default(value: UUID): UuidValidationRuleList {
+        defaultValue = value
+        return this
+    }
+
+    override fun setDefault() {
+        kProperty.set(defaultValue)
+    }
+
+    override fun isOptional() = false
 
 }